cargo version bump works
