+++ releaser/Cargo.toml
@@ -13,3 +13,4 @@ semver = { version = "1.0.23", features = ["serde"] }
 serde = { version = "1.0.210", features = ["derive"] }
 serde_json = "1.0.128"
 toml = "0.8.19"
+toml_edit = "0.22.21"
